refactor: moving login to service
